Added functionality to delete books from list
Modified renderBook() function to create a new button for each book object that allows the user to delete the book from the list. The button sends a DELETE request to the server to delete the corresponding book object. Once the request is successful, the corresponding li element is removed from the DOM.